requirement.txt
snowflake-connector-python==3.10.0
boto3==1.4.8